Added methods

+++ logisimProject/src/controller/JsonFile.java
@@ -1,10 +1,10 @@
 package controller;
 
 public class JsonFile {
-	private final int id;
+	private final long id;
     private final boolean valid;
 
-    public JsonFile(int id, boolean valid) {
+    public JsonFile(long id, boolean valid) {
         this.id = id;
         this.valid = valid;
     }
